# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QA Playground REST API",
    "version": "1.0.0",
    "description": "REST API designed for practicing manual, API, automation, and performance testing. Includes comprehensive error handling, JWT auth with refresh tokens, role-based access control, cart, products, and order management.",
    "contact": {
      "name": "QA Playground Team",
      "email": "support@qaplayground.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/auth/login or /api/auth/register"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Admin User"
          },
          "email": {
            "type": "string",
            "example": "admin@test.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "example": "admin"
          },
          "created_at": {
            "type": "string",
            "example": "2026-10-02 04:08:24"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-10-02 04:08:24"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "List all users (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get user by ID (Owner or Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Access denied (not owner or admin)"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update user profile (Owner or Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "updated@test.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email already in use"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      },
      "delete": {
        "summary": "Delete user by ID (Admin only, cannot delete self)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully (no content)"
          },
          "400": {
            "description": "Cannot delete own account"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  }
}